:recycle:  refactor: styled component style.ts 파일로 분리
